--- numalogic/numaflow/__init__.py
@@ -1,3 +1,0 @@
-from numalogic.numaflow._base import NumalogicUDF
-
-__all__ = ["NumalogicUDF"]
