+++ numalogic/udfs/__init__.py
@@ -1,0 +1,3 @@
+from numalogic.udfs._base import NumalogicUDF
+
+__all__ = ["NumalogicUDF"]
